Fixes after PR review.
