+++ dotnet-base-libraries/ReturnInsertedIdentityWithDapper/ReturnInsertedIdentityWithDapper/Methods/Methods.cs
@@ -3,7 +3,7 @@ using Microsoft.Data.SqlClient;
 
 namespace ReturnInsertedIdentityWithDapper;
 
-public partial class Methods
+public class Methods
 {
     public static int UseOfOutputInserted(string connectionString, Student newStudent)
     {
@@ -15,6 +15,23 @@ public partial class Methods
             $"INSERT INTO Students (Firstname, Surname) " +
             $"OUTPUT INSERTED.Id " +
             $"VALUES ('{newStudent.Firstname}', '{newStudent.Surname}'); ",
+            newStudent
+        );
+
+        dbConnection.Close();
+
+        return lastInsertedId;
+    }
+
+    public static int UseOfScopeIdentity(string connectionString, Student newStudent)
+    {
+        using var dbConnection = new SqlConnection(connectionString);
+
+        dbConnection.Open();
+
+        var lastInsertedId = dbConnection.QuerySingle<int>(
+            $"INSERT INTO Students (Firstname, Surname) VALUES ('{newStudent.Firstname}', '{newStudent.Surname}'); " +
+            "SELECT CAST(SCOPE_IDENTITY() AS INT)",
             newStudent
         );
