--- dotnet-base-libraries/ReturnInsertedIdentityWithDapper/ReturnInsertedIdentityWithDapper/Methods/UseOfScopeIdentity.cs
@@ -1,24 +1,0 @@
-﻿using Dapper;
-using Microsoft.Data.SqlClient;
-
-namespace ReturnInsertedIdentityWithDapper;
-
-public partial class Methods
-{
-    public static int UseOfScopeIdentity(string connectionString, Student newStudent)
-    {
-        using var dbConnection = new SqlConnection(connectionString);
-
-        dbConnection.Open();
-
-        var lastInsertedId = dbConnection.QuerySingle<int>(
-            $"INSERT INTO Students (Firstname, Surname) VALUES ('{newStudent.Firstname}', '{newStudent.Surname}'); " +
-            "SELECT CAST(SCOPE_IDENTITY() AS INT)",
-            newStudent
-        );
-
-        dbConnection.Close();
-
-        return lastInsertedId;
-    }
-}
