+++ dotnet-base-libraries/ReturnInsertedIdentityWithDapper/ReturnInsertedIdentityWithDapper/SetUp.cs
@@ -3,7 +3,7 @@ using Microsoft.Data.SqlClient;
 
 namespace ReturnInsertedIdentityWithDapper;
 
-public class SetUp
+public static class SetUp
 {
     public static void CreateTable(string connectionString)
     {
@@ -11,7 +11,9 @@ public class SetUp
 
         connection.Open();
 
-        var createTableSql = "CREATE TABLE Students (Id INT IDENTITY(1,1) PRIMARY KEY, Surname VARCHAR(255), Firstname VARCHAR(255))";
+        var createTableSql = "CREATE TABLE Students (Id INT IDENTITY(1,1) PRIMARY KEY, " +
+            "Surname VARCHAR(255), " +
+            "Firstname VARCHAR(255))";
 
         connection.Execute(createTableSql);
