+++ dotnet-base-libraries/ReturnInsertedIdentityWithDapper/ReturnInsertedIdentityWithDapper/Student.cs
@@ -2,6 +2,6 @@
 
 public class Student
 {
-    public string? Surname { get; set; }
-    public string? Firstname { get; set; }
+    public string Surname { get; set; }
+    public string Firstname { get; set; }
 }
